pyproject.toml
[project]
name = "mcp-crash-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain-mcp-adapters>=0.1.8",
    "langchain-openai>=0.3.27",
    "langchain[openai]>=0.3.26",
    "langgraph>=0.5.1",
    "python-dotenv>=1.1.1",
]
